# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SuiSaga API
  description: |
    SuiSaga is a blockchain-powered living world game API that provides endpoints for
    authentication, character management, AI-driven dialogue generation, blockchain action
    recording & verification, and demo reliability systems.

    ## Features
    - **Authentication**: Wallet-based JWT authentication
    - **Character Management**: Create, update, and manage characters with memory systems
    - **AI Dialogue**: Generate dynamic dialogue using Z.ai integration
    - **Blockchain Actions**: Record and verify actions on Walrus blockchain
    - **Demo Reliability**: Fallback systems ensuring hackathon demo success

    ## Authentication
    All API endpoints (except health and demo endpoints) require JWT authentication using
    wallet signatures. See the Authentication section for the complete flow.

    ## Rate Limiting
    - Authentication: 10 requests/minute per IP
    - Character API: 100 requests/minute per user
    - Dialogue API: 50 requests/minute per user
    - Actions API: 200 requests/minute per user
  version: 1.0.0
  contact:
    name: SuiSaga Team
    email: team@suisaga.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.suisaga.io
    description: Production server
paths:
...
  /api/dialogue/generate:
    post:
      tags:
        - Dialogue Generation
      summary: Generate AI dialogue
      description: Generates AI-driven dialogue for a character based on the prompt and context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateDialogueRequest'
      responses:
        '200':
          description: Dialogue generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateDialogueResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/dialogue/suggestions:
    post:
      tags:
        - Dialogue Generation
      summary: Get dialogue suggestions
      description: Returns suggested dialogue options based on the context and character.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DialogueSuggestionsRequest'
      responses:
        '200':
          description: Suggestions generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogueSuggestionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/dialogue/history/{characterId}/{playerId}:
    get:
      tags:
        - Dialogue Generation
      summary: Get dialogue history
      description: Retrieves the conversation history between a character and player.
      parameters:
        - $ref: '#/components/parameters/CharacterId'
        - $ref: '#/components/parameters/PlayerId'
      responses:
        '200':
          description: Dialogue history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogueHistoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/authenticate
  schemas:
...
    GenerateDialogueRequest:
      type: object
      required:
        - playerId
        - characterId
        - prompt
      properties:
        playerId:
          type: string
        characterId:
          type: string
        prompt:
          type: string
          maxLength: 500
        context:
          type: string
          maxLength: 200
    GenerateDialogueResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            dialogue:
              type: string
            characterId:
              type: string
            emotionalTone:
              type: string
            consequences:
              type: array
              items:
                $ref: '#/components/schemas/Consequence'
    Consequence:
      type: object
      properties:
        system:
          type: string
        narrative:
          type: string
        severity:
          type: string
          enum: [minor, major, critical]
        impact:
          type: integer
          minimum: 1
          maximum: 10
...
  parameters:
    CharacterId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Unique character identifier
      example: char_12345
...
    PlayerId:
      name: playerId
      in: path
      required: true
      schema:
        type: string
      description: Unique player identifier
      example: player_12345
...
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: VALIDATION_ERROR
              message:
                type: string
                example: Invalid request parameters
              details:
                type: array
                items:
                  type: string
              metadata:
                $ref: '#/components/schemas/ErrorMetadata'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: NOT_FOUND
              message:
                type: string
                example: The requested resource was not found
              metadata:
                $ref: '#/components/schemas/ErrorMetadata'
...